api/requirements.txt
aioice==0.9.0
aiortc==1.10.1
annotated-types==0.7.0
anyio==4.8.0
attrs==22.2.0
av==12.3.0
bcrypt==4.2.1
Brotli==1.0.9
certifi==2022.9.24
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.8
colorzero==2.0
crit==3.17.1
cryptography==44.0.1
distro==1.8.0
dnspython==2.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
google-crc32c==1.6.0
gpiozero==2.0.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.3
ifaddr==0.2.0
influxdb_client==1.48.0
Jinja2==3.1.5
jsonpointer==2.3
jsonschema==4.10.3
lgpio==0.2.2.0
libarchive-c==2.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mutagen==1.47.0
numpy==1.24.2
olefile==0.46
opencv-python==4.11.0.86
passlib==1.7.4
picamera2==0.3.25
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
Pillow==9.4.0
protobuf==4.21.12
PyAudio==0.2.14
pycparser==2.22
pycryptodomex==3.11.0
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.11.0
PyOpenGL==3.1.6
pyOpenSSL==25.0.0
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyrsistent==0.18.1
python-apt==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-multipart==0.0.20
python-prctl==1.8.1
PyYAML==6.0
reactivex==4.0.4
requests==2.28.1
rfc3987==1.3.8
rich==13.9.4
rich-toolkit==0.13.2
RPi.GPIO==0.7.1
shellingham==1.5.4
simplejpeg==1.6.6
six==1.16.0
smbus2==0.4.2
sniffio==1.3.1
spidev==3.5
ssh-import-id==5.10
starlette==0.45.3
texttable==1.6.7
toml==0.10.2
tqdm==4.64.1
typer==0.15.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.34.0
uvloop==0.21.0
v4l2-python3==0.3.5
watchfiles==1.0.4
webcolors==1.11.1
websocket-client==1.2.3
websockets==15.0
